==> 02_Exploratory_Data_Analysis/01_EDA_AND_DATA_CLEAN.ipynb <==
### EXPLORAÇÃO DE DADOS - GE - Cursos tecnicos

!dir

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

df =  pd.read_csv("df_2016_ate_2019.csv", encoding='utf8',  error_bad_lines=False, sep=";")

colunas_ = colunas.map(lambda x: x.replace(" ", "_"))
colunas_ = colunas_.map(lambda x: x.replace("ó", "o"))
colunas_ = colunas_.map(lambda x: x.replace("Á", "A"))
colunas_ = colunas_.map(lambda x: x.replace("é", "e"))
colunas_ = colunas_.map(lambda x: x.replace("ó", "o"))
colunas_ = colunas_.map(lambda x: x.replace("í", "i"))
colunas_ = colunas_.map(lambda x: x.replace("ã", "a"))
colunas_ = colunas_.map(lambda x: x.replace("á", "a"))
colunas_ = colunas_.map(lambda x: x.replace("ç", "c"))
colunas_ = colunas_.map(lambda x: x.replace(".", ""))

df.columns = colunas_

 for curso in df.Nome_Curso.unique(): print(curso)

df_automacao = df.loc[(df.Nome_Curso == "Técnico em Automação Industrial")]

len(df_automacao.columns)

for i, col in enumerate(df_automacao.columns): print(col, i)

### Selecionar todas as linhas de 20  em 20 colunas

df_automacao_01 = df_automacao.iloc[:, 0:20]
df_automacao_02 = df_automacao.iloc[:, 20:40]
df_automacao_03 = df_automacao.iloc[:, 40:60]
df_automacao_04 = df_automacao.iloc[:, 60:80]
df_automacao_05 = df_automacao.iloc[:, 80:100]
df_automacao_06= df_automacao.iloc[:, 100:]

### Selecionando dados para proxima fase

##### PRIMEIRO DATAFRAME SELECIONADO

dataframe_01 = df_automacao_01[['Periodo_Letivo','Matriz_Aplicada', 'Cod_Turma', 'Nome_Turma']].copy()

dataframe_01.isnull().sum()

dataframe_01.info()

----------------------------

#### SEGUNDO DATAFRAME SELECIONADO

dataframe_02 = df_automacao_02[['Data_Inicio_Turma', 'Nome_Aluno','Data_Fim_Turma', 'Raca', 'Idade', 'Nacionalidade', 'Sexo', 'Nome_Aluno', 'Cod_Pessoa']].copy()

dataframe_02.Data_Fim_Turma = pd.to_datetime(dataframe_02.Data_Fim_Turma)
dataframe_02.Data_Inicio_Turma = pd.to_datetime(dataframe_02.Data_Inicio_Turma)

dataframe_02.Raca = dataframe_02.Raca.replace(np.nan, 'Não informado')
dataframe_02.Raca = dataframe_02.Raca.replace('Não declarado', 'Não informado')

dataframe_02.isnull().sum()

dataframe_02.info()

----------------------------------------------

### Terceiro Dataframe Selecionado

dataframe_03 = df_automacao_03[['Cidade', 'Cidade', 'Grau_Instrucao', 'Profissao', 'Deficiências', 'Tipo_Aluno']].copy()

dataframe_03.Profissao = dataframe_03.Profissao.replace(np.nan, "naoinformado")
dataframe_03["Deficiências"] = dataframe_03["Deficiências"].replace(np.nan, "naoapresenta")

dataframe_03.Tipo_Aluno.unique()

dataframe_03.Tipo_Aluno = dataframe_03.Tipo_Aluno.replace(np.nan, 'Desconhecido')

dataframe_03.isnull().sum()

dataframe_03.info()

### Quarto Dataframe Selecionado

dataframe_04 =  df_automacao_04[['Periodo_Matricula', 'Status_Matricula', 'Resultado_Final', 'Resultado_Hab_Aluno', 'Data_Matricula', 'Data_Criacao_Matricula',
       'Data_Prevista_Saida','Disciplina', 'Data_Inicio_Disciplina',
       'Data_Fim_Disciplina', 'Media_Av_1', 'Media_Av_2', 'Recuperacao',
       'Media_Final', 'Total_de_Faltas']].copy()

dataframe_04.drop('Resultado_Final', axis=1, inplace=True)

dataframe_04.drop('Recuperacao', axis=1, inplace=True)

dataframe_04.Total_de_Faltas = dataframe_04.Total_de_Faltas.replace(np.nan, 0.0)

dataframe_04.Media_Av_1 = dataframe_04.Media_Av_1.replace(np.nan, 0.0)

dataframe_04.Media_Av_2 = dataframe_04.Media_Av_2.replace(np.nan, 0.0)

dataframe_04.Media_Final = dataframe_04.Media_Final.replace(np.nan, 0.0)

dataframe_04.isna().sum()

#dataframe_04.dropna(inplace=True)

dataframe_04.Data_Prevista_Saida = pd.to_datetime(dataframe_04.Data_Prevista_Saida)

dataframe_04.info()

--------------------------

### Quinto Dataframe Selecionado

dataframe_05 = df_automacao_05[['Condicao_Aluno', 'Resp_Financeiro']].copy()

dataframe_05.Condicao_Aluno = dataframe_05.Condicao_Aluno.replace(np.nan, 'Nao_Informado')
dataframe_05.Resp_Financeiro = dataframe_05.Resp_Financeiro.replace(np.nan, 'Nao_Informado')

dataframe_05.info()

-----------------------------

### Inicialmente nenhum dado selecionado do df_automacao_06

# COCATENANDO 

dataframe_final = pd.concat([dataframe_01, dataframe_02, dataframe_03, dataframe_04, dataframe_05], axis=1)

-------------------------------------
---------------------------------

dataframe_final.Cod_Pessoa.nunique()

dataframe_final.Cod_Pessoa.loc[dataframe_final.Periodo_Letivo == '2017/1'].nunique()

alunos_2017_1 = dataframe_final.Cod_Pessoa.loc[dataframe_final.Periodo_Letivo == '2017/1']

dataframe_final.loc[dataframe_final.Cod_Pessoa == 814280]

lista_de_notas = ['FdC_Media_Final', 'FdE_Media_Final',
       'FdM_Media_Final', 'ADA_Media_Final',
       'PdS_Media_Final']

### As medias vao funcionar para qual semestre ???  
POC 3 Semestre  Notas dos alunos ate o 2 semestre e depois novamente para o 3.

Se eu treinar um algoritmo com as notas de todos os semestres. Eu teria que fornecer na hora da predicao usuarios que tb fizeram todos os semestres do algoritmo de treino. O que nao faz sentido.

for id_aluno in alunos_2017_1:
    dataframe_final.loc[(dataframe_final.Cod_Pessoa==id_aluno) ,'FdC_Media_Final'] = dataframe_final.loc[(dataframe_final.Disciplina == 'Fundamentos da Comunicação') & (dataframe_final.Cod_Pessoa == id_aluno)].Media_Final
    dataframe_final.loc[(dataframe_final.Cod_Pessoa==id_aluno),'FdE_Media_Final'] = dataframe_final.loc[(dataframe_final.Disciplina == 'Fundamentos da Eletrotécnica') & (dataframe_final.Cod_Pessoa == id_aluno)].Media_Final
    dataframe_final.loc[(dataframe_final.Cod_Pessoa==id_aluno),'FdM_Media_Final'] = dataframe_final.loc[(dataframe_final.Disciplina == 'Fundamentos da Mecânica') & (dataframe_final.Cod_Pessoa == id_aluno)].Media_Final
    dataframe_final.loc[(dataframe_final.Cod_Pessoa==id_aluno),'ADA_Media_Final'] = dataframe_final.loc[(dataframe_final.Disciplina == 'Acionamento de Dispositivos Atuadores') & (dataframe_final.Cod_Pessoa == id_aluno)].Media_Final
    dataframe_final.loc[(dataframe_final.Cod_Pessoa==id_aluno),'PdS_Media_Final'] = dataframe_final.loc[(dataframe_final.Disciplina == 'Processamento de Sinais') & (dataframe_final.Cod_Pessoa == id_aluno)].Media_Final
    df_sample = dataframe_final.loc[dataframe_final.Cod_Pessoa==id_aluno].copy()
    df_sample.FdC_Media_Final = df_sample.FdC_Media_Final.sum()
    df_sample.FdE_Media_Final = df_sample.FdE_Media_Final.sum()
    df_sample.FdM_Media_Final = df_sample.FdM_Media_Final.sum()
    df_sample.ADA_Media_Final = df_sample.ADA_Media_Final.sum()
    df_sample.PdS_Media_Final = df_sample.PdS_Media_Final.sum()
    dataframe_final.update(df_sample)

dataframe_final.dropna(inplace=True)

dataframe_final.drop('Matriz_Aplicada', inplace=True, axis=1)

dataframe_final.drop(['Periodo_Letivo','Cod_Turma', 'Nome_Turma', 'Data_Inicio_Turma' , 'Data_Fim_Turma'], inplace=True, axis=1)

dataframe_final.drop(['Cod_Pessoa', 'Data_Criacao_Matricula', 'Data_Prevista_Saida'], inplace=True, axis=1)

dataframe_final.drop(['Cidade'], inplace=True, axis=1)

dataframe_final.drop(['Data_Matricula'], inplace=True, axis=1)

pd.get_dummies(dataframe_final.Raca, prefix='Raca')

dataframe_final.iloc[:, 0:20]

dataframe_final.iloc[:, 20:]

dataframe_final.to_csv('dataframe_completo.csv', sep=';')

---------------------------------------------------
