==> 02_Exploratory_Data_Analysis/02_DATA_WRANGLING.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv('dataframe_completo.csv', sep=';')

df.drop('Unnamed: 0', axis=1, inplace=True)

df['Comp_Financeiro'] = df.Nome_Aluno == df.Resp_Financeiro

df.loc[df.Comp_Financeiro == False].head(2)

df.drop(['Media_Av_1', 'Media_Av_2', 'Media_Final'], inplace=True , axis=1)

df.Comp_Financeiro = df.Comp_Financeiro.astype(int)

df.drop(['Resp_Financeiro'], inplace=True , axis=1)

df.drop(['Nome_Aluno.1'], inplace=True , axis=1)

raca = pd.get_dummies(df.Raca, prefix="Raca_")

df.drop(['Raca'], inplace=True , axis=1)

nacionalidade = pd.get_dummies(df.Nacionalidade, prefix="Nacionalidade_")

df.drop(['Nacionalidade'], inplace=True , axis=1)

grau = pd.get_dummies(df.Grau_Instrucao, prefix="Grau_")

df.drop(['Grau_Instrucao'], inplace=True , axis=1)

sexo = pd.get_dummies(df.Sexo, prefix="Sexo_")

df.drop(['Sexo'], inplace=True , axis=1)

profissao = pd.get_dummies(df.Profissao, prefix="Profissao_")

deficiencias = pd.get_dummies(df['Deficiências'], prefix="Deficiencias_")

tipo = pd.get_dummies(df.Tipo_Aluno, prefix="Tipo_")

df.drop(['Deficiências', 'Profissao', 'Tipo_Aluno'], inplace=True , axis=1)

df.drop(['Data_Inicio_Disciplina', 'Data_Fim_Disciplina', 'Disciplina'], inplace=True, axis=1)

condicao = pd.get_dummies(df.Condicao_Aluno, prefix="Condicao_")

df.drop(['Condicao_Aluno'], inplace=True , axis=1)

status = pd.get_dummies(df.Status_Matricula, prefix="Status_")

df.drop(['Status_Matricula'], inplace=True , axis=1)

label = pd.get_dummies(df.Resultado_Hab_Aluno, prefix="Label_")

df.drop(['Resultado_Hab_Aluno'], inplace=True , axis=1)

df.drop(['Periodo_Matricula'], inplace=True , axis=1)
df.drop(['Total_de_Faltas'], inplace=True , axis=1)

dataframe_automacao = pd.concat([df, raca, nacionalidade, grau, sexo, profissao, deficiencias, tipo, condicao, status, label],axis=1)

dataframe_automacao.set_index('Nome_Aluno', inplace=True)
dataframe_automacao

dataframe_automacao.drop_duplicates(inplace=True)

dataframe_automacao.drop(['Status__Matriculado', 'Status__Pré-Matricula'], inplace=True, axis=1)

dataframe_automacao.drop(['Status__Cancelado', 'Status__Desistente', 'Status__Evadido'], inplace=True, axis=1)

dataframe_automacao.drop(['Status__Trancado'], inplace=True, axis=1)

dataframe_automacao.drop(['Tipo__Ativo', 'Tipo__Desconhecido',
       'Tipo__Inativo'], inplace=True, axis=1)

Cancelado
Concluinte
Desistente
Evadido
Matriculado
Reprovado
Trancado

df_concluintes = dataframe_automacao.loc[(dataframe_automacao.Label__Concluinte > 0)].copy()
df_concluintes.shape

df_evadidos = dataframe_automacao.loc[(dataframe_automacao.Label__Desistente > 0)|(dataframe_automacao.Label__Cancelado > 0)|(dataframe_automacao.Label__Evadido > 0)].copy()
df_evadidos.shape

df_concluintes.drop(['Label__Cancelado', 'Label__Concluinte', 'Label__Desistente',
       'Label__Evadido', 'Label__Matriculado', 'Label__Reprovado',
       'Label__Trancado'], inplace=True, axis=1)

df_evadidos.drop(['Label__Cancelado', 'Label__Concluinte', 'Label__Desistente',
       'Label__Evadido', 'Label__Matriculado', 'Label__Reprovado',
       'Label__Trancado'], inplace=True, axis=1)

df_concluintes['Label'] = 1
df_evadidos['Label'] = 0

dataframe_final = pd.concat([df_concluintes, df_evadidos], axis=0)

dataframe_final.to_csv('df_automacao.csv', sep=';')
