==> 02_Exploratory_Data_Analysis/03_FEATURES_ENGINNERING.ipynb <==
import pandas as pd
import numpy as np
from sklearn import preprocessing

!dir

df = pd.read_csv('df_automacao.csv', sep=';')
df.shape

alunos = df.Nome_Aluno

df_completo = pd.read_csv('dataframe_completo.csv', sep=';')

df_completo.loc[df_completo.Nome_Aluno == "Abel Giachini Frare"].Total_de_Faltas.sum()

subset = df_completo[['Nome_Aluno', 'Total_de_Faltas']]

datagrouped = subset.groupby(['Nome_Aluno']).agg(Faltas=("Total_de_Faltas","sum"))

datagrouped.reset_index(inplace=True)

mapper = dict(zip(datagrouped.Nome_Aluno, datagrouped.Faltas))
df["Total_de_Faltas"] = df.loc[:,"Total_de_Faltas"] = df.Nome_Aluno.map(mapper)

alunos = df.Nome_Aluno.unique()

x = df[['Idade', 'FdC_Media_Final', 'FdE_Media_Final',
       'FdM_Media_Final', 'ADA_Media_Final', 'PdS_Media_Final','Total_de_Faltas']] #returns a numpy array

min_max_scaler = preprocessing.MinMaxScaler()
x_scaled = min_max_scaler.fit_transform(x)
x_scaled = pd.DataFrame(x_scaled, columns=['Idade', 'FdC_Media_Final', 'FdE_Media_Final',
       'FdM_Media_Final', 'ADA_Media_Final', 'PdS_Media_Final','Total_de_Faltas'])

df.update(x_scaled)


df_f = df[['Nome_Aluno', 'Idade', 'FdC_Media_Final', 'FdE_Media_Final',
       'FdM_Media_Final', 'ADA_Media_Final',
       'Total_de_Faltas','PdS_Media_Final',
       'Comp_Financeiro', 'Raca__Amarela', 'Raca__Branca',
       'Raca__Não informado', 'Raca__Parda', 'Raca__Preta',
       'Nacionalidade__Brasileira', 'Nacionalidade__Paraguaia',
       'Grau__Educação superior completo',
       'Grau__Educação superior incompleto', 'Grau__Ensino médio completo',
       'Grau__Ensino médio incompleto', 'Grau__Pós Grad. completo',
       'Sexo__Feminino', 'Sexo__Masculino', 'Profissao__Secretário',
       'Profissao__naoinformado',
       'Deficiencias__Deficiência Física (Amputação)',
       'Deficiencias__naoapresenta',
       'Condicao__Pago por Pessoa Fisica ou Empresa', 'Label']].copy()

df_f.to_csv('dataframe_automacao_final_tf.csv', sep=';')
